layout/base/tests/test_bug1550869_video.html

<!DOCTYPE HTML>
<html>
  <title>Bug 1516963: Test AccessibleCaret doesn't show when clicking on an empty video container.</title>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <style>
  #container {
    border: 1px solid blue;
    width: 100px;
    height: 100px;
    background-color: yellow;
  }
  </style>
  <script>
  SimpleTest.waitForExplicitFinish();

  document.addEventListener("selectionchange", () => {
    ok(window.getSelection().isCollapsed, "The selection should be collapsed!");
  });

  function click() {
    ok(window.getSelection().isCollapsed, "The selection should start collapsed!");
    let container = document.getElementById("container");
    synthesizeMouseAtCenter(container, {});
    setTimeout(() => {
      ok(window.getSelection().isCollapsed, "The selection should remain collapsed!");
      SimpleTest.finish();
    });
  }
  </script>
  <body onload="SimpleTest.waitForFocus(click);">
    <video id="container"></video>
  </body>
</html>
